+++ plugins/hello-world/webapp/view/jpulse-common.js
@@ -3,8 +3,8 @@
  * @tagline         Common JavaScript of the Hello World Plugin
  * @description     Common JavaScript of the Hello World Plugin, appended to the framework common JavaScript
  * @file            plugins/hello-world/webapp/view/jpulse-common.js
- * @version         1.6.2
- * @release         2026-01-30
+ * @version         1.6.3
+ * @release         2026-01-31
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/