+++ templates/deploy/ecosystem.dev.config.cjs
@@ -7,8 +7,8 @@
  * @site            %SITE_NAME%
  * @generated       %GENERATION_DATE%
  * @file            templates/deploy/ecosystem.dev.config.cjs
- * @version         1.6.2
- * @release         2026-01-30
+ * @version         1.6.3
+ * @release         2026-01-31
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/