+++ webapp/tests/integration/w047-site-files.test.js
@@ -3,8 +3,8 @@
  * @tagline         Integration tests for W-047 site-specific file loading (W-098 append mode)
  * @description     Tests jpulse-common.css/js append mode and site-specific templates
  * @file            webapp/tests/integration/w047-site-files.test.js
- * @version         1.6.2
- * @release         2026-01-30
+ * @version         1.6.3
+ * @release         2026-01-31
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/