+++ webapp/tests/unit/controller/handlebar-plugin-interface.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for W-116: Handlebars helper plugin interface
  * @description     Tests for custom helper registration, auto-discovery, and JSDoc extraction
  * @file            webapp/tests/unit/controller/handlebar-plugin-interface.test.js
- * @version         1.6.2
- * @release         2026-01-30
+ * @version         1.6.3
+ * @release         2026-01-31
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/